requirements.txt
telethon
questionary
rich
python-dotenv
